StegoNet/Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.1
torch==2.7.0
torchvision==0.22.0
Pillow==11.1.0